[DO NOT MERGE] Added Star and Hybrid
Added the star and hybrid formulations. Lots of duplicate code, but
things are working end 2 end.

The precision of the timer is annoyingly low, would like to get more
precision if possible.

Would also like to clean up this code into a class so things can be a
bit cleaner.