[DO NOT MERGE] Added Sparsity
Made the clique, star, and hybrid models directly sparse. They no longer
create a dense matrix before getting the sparse view; they directly
generate the sparse matrix. Not sure if this is the most efficient way
to do this, but it was simple to do and it did reduce the memory
footprint. Timing was not effected for execution.