[VPR][CMake] Added Sub-Libraries to VPR
VPR has become such a large project that incremental changes to one of
its sub-features yields very long compile times. There are also many
different dependencies needed for only one sub-feature which is included
everywhere which leads to lots of polution and bloat.

By creating libraries for each of the sub-folders of VPR, it should
provide the ability to clean some of this up.