[Parallel Router] Simplified Heap Node and Rewrote Routing Algorithm
Stored only a small amount of information (i.e., node ID) in the heap to
keep it lightweight, and rewrote the routing algorithm such that the
global updates happen just before pushing a neighbor node to the heap.

This commit is not workable. It cannot successfully routes any circuit
in the MCNC benchmark.