[WIP] Fixed Infinite Loop Bug
This fixes the infinite loop found in the bitcoin_miner circuit.

This changes the backward and total cost calculations from float to
double. This also changes the heap such that it stores double instead of
float. The heap may be able to be turned into float; but, it requires
some unsavory code.

The total cost may also be able to remain a float; however, need to see
if master would change this to a double.

Also left in some debug code which prevents infinite loops so we do not
crash the testing machines. This can be removed once everything works.