[Route] Increased the Precision of Cost Terms
For very large circuits, it has been found that the cost terms grow to
such a degree that adding the (considerably smaller) edge cost terms had
no effect. This can cause problems for algorithms that require the
accuracy and may yield lower quality of results.

Changed the cost terms from float to double.