[ParallelRouter] Fixed Bitcoin Bug
The bug in bitcoin_miner was caused by a precision issue where the
congestion cost would get so large that the delay cost would not matter.
This creates zero-cost cycles which the deterministic parallel router
cannot handle.

There was a PR on VTR Master which fixed this by preventing the pres_fac
from going over 1000. Bringing that change in early. This change will
fully be brought in the next time this branch is rebased onto master.