[ParallelRouter] Reverted Changes to Serial Connection Router
At the start of the project we cleaned up the Serial Connection Router
(removing things that were not necessary for function and getting things
ready to be parallelized). This was used as the basis of the Parallel
Connection Router.

This is no longer being used, so to make it easier to rebase onto
master; this is being removed.