[ClusterLegalizer] Cluster Legalizer API
Isolated the cluster legalization logic out of the packer's clustering
algorithm in such a way that it can be used externally to the clusterer
with no issue. Once the ClusterLegalizer API is constructed, it will
allocate all of the memory that it needs for legalization; it can then
be used to legalize clusters; and when it is destroyed it will remove
all state created for clustering.

The cluster legalizer API maintains the legalized clusters internally as
opposed to using the ClusterBlockId as an ID to the clusters. This
allowed me to separate the ClusteredNetlist from the packing algorithm
entirely. This importantly helper separate out all legalization logic
from directly modifying the ClusteredNetlist generation code.

I did my best to remove as much global state as I could from the API to
make it as self-contained as possible.

This change greatly cleaned up the packer code and removed many of its
global variables (including the ClusterLegalizer itself, which is now a
local variable!)

This change also removed the re_clustering API since it would need to be
heavily modified to include this change; and since this feature was not
being tested, it would be challenging to upgrade to use this new cluster
legalization API.