[ClusterLegalizer] Moved Stat Counting Into Cluster Legalizer
Found that one of the pb_stats members is required for cluster
legalization; but was being calculated outside of the cluster legalizer.

Moved this in to allow the cluster legalizer to be used outside of the
clusterer.

Also found an issue where two clusters of the same type cannot be
constructed at the same time. Tried fixing it, but it produces different
results (different clusters). Will raise in a separate PR, left as a
fixme comment.