[Place] Moved PlaceMacros Out of BlkLocRegistry
The PlaceMacros only require the Clustering and atom/arch info to be
constructed; after which they are never modified. Since they are
unchanged by the placement, they were out of place in the
BlkLocRegistry.

Moved them to the clustering context since they only make sense once the
clusters are loaded.

This will make it easier to separate out the initial placement from the
SA placer.